Refactor: Break down PhoneSignUpForm component
Refactor the PhoneSignUpForm component into smaller, more manageable sub-components to improve readability and maintainability.
